+++ src/main/java/com/verdantartifice/thaumicwonders/common/tiles/devices/TileDimensionalRipper.java
@@ -1,6 +1,7 @@
 package com.verdantartifice.thaumicwonders.common.tiles.devices;
 
 import java.awt.Color;
+import java.util.List;
 
 import com.verdantartifice.thaumicwonders.ThaumicWonders;
 import com.verdantartifice.thaumicwonders.common.blocks.BlocksTW;
@@ -313,15 +314,24 @@ public class TileDimensionalRipper extends TileTW implements IAspectContainer, I
     }
 
     protected void createRift(BlockPos pos, int fuelUsed) {
-        if (EntityUtils.getEntitiesInRange(world, pos, null, EntityFluxRift.class, 32.0D).size() > 0) {
-            return;
-        }
-        EntityFluxRift rift = new EntityFluxRift(this.world);
-        rift.setRiftSeed(this.world.rand.nextInt());
-        rift.setLocationAndAngles(pos.getX() + 0.5D, pos.getY() + 0.5D, pos.getZ() + 0.5D, (float)this.world.rand.nextInt(360), 0.0F);
-        double size = Math.sqrt((2 * fuelUsed) * 3.0F);
-        if (this.world.spawnEntity(rift)) {
-            rift.setRiftSize((int)size);
+        List<EntityFluxRift> localRifts = EntityUtils.getEntitiesInRange(this.world, pos, null, EntityFluxRift.class, 2.0D);
+        if (localRifts.size() > 0) {
+            // Enlarge target rift
+            EntityFluxRift rift = localRifts.get(0);
+            int oldSize = rift.getRiftSize();
+            double oldFuel = (oldSize * oldSize) / 3.0D;
+            int newSize = (int)Math.sqrt((oldFuel + (2 * fuelUsed)) * 3.0D);
+            rift.setRiftSize(newSize);
+            rift.setRiftStability(rift.getRiftStability() - (newSize - oldSize));
+        } else if (EntityUtils.getEntitiesInRange(this.world, pos, null, EntityFluxRift.class, 32.0D).size() == 0) {
+            // Create new rift if no others are nearby
+            EntityFluxRift rift = new EntityFluxRift(this.world);
+            rift.setRiftSeed(this.world.rand.nextInt());
+            rift.setLocationAndAngles(pos.getX() + 0.5D, pos.getY() + 0.5D, pos.getZ() + 0.5D, (float)this.world.rand.nextInt(360), 0.0F);
+            double size = Math.sqrt((2 * fuelUsed) * 3.0D);
+            if (this.world.spawnEntity(rift)) {
+                rift.setRiftSize((int)size);
+            }
         }
     }